Crew Part Controller, View 정리 각각의 클래스들을 crew package를 추가하여 넣어서 정리 part가 많아서 헷갈릴 수 있어서
