+++ miniProject/src/com/kh/model/vo/Crew.java
@@ -7,7 +7,7 @@ public class Crew {
 	private String crewName; // 크루명
 	private User crewMaster; // 크루장
 	private ArrayList<User> userList; // 크루원 리스트
-	private boolean isAccept; // 크루 만들기 승인 여부 관리자가 승인
+	private boolean isAccept; // 크루 만들기 승인 여부 관리자가 승인 기본값 false
 
 	public Crew() {
 		userList = new ArrayList<User>();
@@ -17,7 +17,7 @@ public class Crew {
 		this();
 		this.crewName = crewName;
 		this.crewMaster = crewMaster;
-		this.isAccept = isAccept;
+//		this.isAccept = false; // 기본값은 승인이 안된 상태 
 	}
 
 	public boolean isAccept() {
