+++ miniProject/src/com/kh/view/crew/CrewRankFrame.java
@@ -1,4 +1,4 @@
-package com.kh.view;
+package com.kh.view.crew;
 
 import java.awt.BorderLayout;
 import java.awt.EventQueue;
@@ -7,20 +7,20 @@ import javax.swing.JFrame;
 import javax.swing.JPanel;
 import javax.swing.border.EmptyBorder;
 
-import com.kh.controller.CrewCreateController;
+import com.kh.controller.CrewRankController;
 import javax.swing.JButton;
 import java.awt.event.ActionListener;
 import java.awt.event.ActionEvent;
 
-public class CrewCreateFrame extends JFrame {
+public class CrewRankFrame extends JFrame {
 
 	private CrewViewManager crewManager;
 	
-	private CrewCreateController createController;
+	private CrewRankController rankController;
 
 	private JPanel contentPane;
 
-	public CrewCreateFrame() {
+	public CrewRankFrame() {
 		setDefaultCloseOperation(JFrame.EXIT_ON_CLOSE);
 		setBounds(100, 100, 360, 600);
 		contentPane = new JPanel();
@@ -28,36 +28,37 @@ public class CrewCreateFrame extends JFrame {
 		setContentPane(contentPane);
 		contentPane.setLayout(null);
 
-		JButton btnNewButton = new JButton("완료");
+		JButton btnNewButton = new JButton("뒤로가기");
 		btnNewButton.addActionListener(new ActionListener() {
 			public void actionPerformed(ActionEvent e) {
-				crewManager.convertFrame("main"); // main page로
+				// 특정 crew 페이지로 다시 이동
+				crewManager.convertFrame("crew");
 			}
 		});
-		btnNewButton.setBounds(235, 20, 97, 23);
+		btnNewButton.setBounds(12, 10, 97, 23);
 		contentPane.add(btnNewButton);
 
 		JButton btnNewButton_1 = new JButton("홈");
 		btnNewButton_1.addActionListener(new ActionListener() {
 			public void actionPerformed(ActionEvent e) {
-				crewManager.convertFrame("main"); // main page로
+				crewManager.convertFrame("main"); // main 페이지로
 			}
 		});
-		btnNewButton_1.setBounds(131, 528, 97, 23);
+		btnNewButton_1.setBounds(125, 528, 97, 23);
 		contentPane.add(btnNewButton_1);
 	}
 
-	public CrewCreateFrame(CrewViewManager crewManager, CrewCreateController createController) {
+	public CrewRankFrame(CrewViewManager crewManager, CrewRankController rankController) {
 		this();
 		this.crewManager = crewManager;
-		this.createController = createController;
+		this.rankController = rankController;
 	}
 
 	public static void main(String[] args) {
 		EventQueue.invokeLater(new Runnable() {
 			public void run() {
 				try {
-					CrewCreateFrame frame = new CrewCreateFrame();
+					CrewRankFrame frame = new CrewRankFrame();
 					frame.setVisible(true);
 				} catch (Exception e) {
 					e.printStackTrace();
@@ -65,4 +66,5 @@ public class CrewCreateFrame extends JFrame {
 			}
 		});
 	}
+
 }
